requirements.txt
pathlib
pandas
openpyxl
python-docx